clipping

+++ prototypes/amr/definitions.h
@@ -6,7 +6,7 @@ typedef float Realf;
 typedef float Real;
 typedef const float creal;
 
-typedef std::array<uint64_t, 3> indices_t;
+typedef std::array<size_t, 3> indices_t;
 
 typedef std::array<Real, 4> vblock_t;
